Replace mpay locking with atomic

+++ server/server.go
@@ -36,7 +36,7 @@ type Info struct {
 	SSLRequired       bool     `json:"ssl_required"` // DEPRECATED: ssl json used for older clients
 	TLSRequired       bool     `json:"tls_required"`
 	TLSVerify         bool     `json:"tls_verify"`
-	MaxPayload        int      `json:"max_payload"`
+	MaxPayload        int64    `json:"max_payload"`
 	IP                string   `json:"ip,omitempty"`
 	ClientConnectURLs []string `json:"connect_urls,omitempty"` // Contains URLs a client can connect to.
